+++ chunky_mem/include/chunky_mem/chunk_static_slots.h
@@ -13,7 +13,7 @@ namespace chunky_mem {
         size_t element_size_t, 
         size_t chunk_size_t = (1024*64 - ((1024*64) % element_size_t)) / element_size_t, 
         size_t alignment_size_t = 16,
-        typename idx_t = int,
+        typename idx_t = size_t,
         typename size_t = size_t
     >
     struct ChunkStaticSlots
