+++ chunky_mem/include/chunky_mem/linked_list.h
@@ -7,7 +7,7 @@
 
 namespace chunky_mem {
 
-    template<typename idx_t = int, typename size_t = idx_t>
+    template<typename idx_t = size_t, typename size_t = idx_t>
     struct LinkedList
     {
         using idx_type = idx_t;
